refactor : challenge 테이블에서 화면을 그리기 위한 용도의 컬럼들(starRating,totalStarRating,reviewCount)를 제거함

+++ challenge-service/src/main/java/com/challengers/challengeservice/userchallenge/repository/UserChallengeRepository.java
@@ -11,7 +11,7 @@ import java.util.Optional;
 
 public interface UserChallengeRepository extends JpaRepository<UserChallenge,Long>, UserChallengeRepositoryCustom {
     Optional<UserChallenge> findByUserIdAndChallengeId(Long userId, Long challengeId);
-    Long countByChallengeId(Long challengeId);
+    int countByChallengeId(Long challengeId);
 
     List<UserChallenge> findByChallengeId(Long challengeId);
